refactor: OAuthProperties 파일 상위 패키지로 이동
